requirements.txt
httpx>=0.27.0
pymongo>=4.7.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
googlesearch-python>=1.2.0
tenacity>=8.2.0
openai>=1.30.0
boto3>=1.34.0
